Actor-based movement system was undeprecated. Border checks for minX_i(), maxX_i(), minY_i(), maxY_i() were added.
Complex aspectRatio-based projection calculations were deprecated and removed.

+++ Engines/Camera/Camera.hpp
@@ -40,10 +40,10 @@ public:
 	float maxX() const;
 	float maxY() const;
 
-	unsigned int minX_i() const;
-	unsigned int minY_i() const;
-	unsigned int maxX_i() const;
-	unsigned int maxY_i() const;
+	size_t minX_i() const;
+	size_t minY_i() const;
+	size_t maxX_i() const;
+	size_t maxY_i() const;
 	Map* map();
 
 	inline bool wasCameraChanged() const {
